+++ clang-tools-extra/clang-tidy/autosar/AutosarTidyModule.cpp
@@ -8,6 +8,7 @@
 
 #include "../ClangTidyModule.h"
 #include "../ClangTidyModuleRegistry.h"
+#include "InterfaceConnectedCheck.h"
 #include "MemoryMappingCheck.h"
 #include "TypeUsageCheck.h"
 
@@ -17,6 +18,8 @@ namespace autosar {
 class AutosarModule : public ClangTidyModule {
 public:
   void addCheckFactories(ClangTidyCheckFactories &CheckFactories) override {
+    CheckFactories.registerCheck<InterfaceConnectedCheck>(
+        "autosar-interface-connected");
     CheckFactories.registerCheck<MemoryMappingCheck>("autosar-memory-mapping");
     CheckFactories.registerCheck<TypeUsageCheck>("autosar-type-usage");
   }